requirements.txt
discord.py
requests
pytz
google-generativeai
